002_CustomMobs : Fixed active td game list not removing inactive by adding equals() in TowerDefense + Moved all TDGame logic to TowerDefenseManager from service
